+++ evs/include/huaweicloud/evs/v2/EvsRegion.h
@@ -22,24 +22,78 @@ public:
 
     static std::map<std::string, Region> initRegions() {
         std::map<std::string, Region> regions;
-        regions["af-south-1"] = Region("af-south-1","https://evs.af-south-1.myhuaweicloud.com") ;
-        regions["cn-north-4"] = Region("cn-north-4","https://evs.cn-north-4.myhuaweicloud.com") ;
-        regions["cn-north-1"] = Region("cn-north-1","https://evs.cn-north-1.myhuaweicloud.com") ;
-        regions["cn-north-2"] = Region("cn-north-2","https://evs.cn-north-2.myhuaweicloud.com") ;
-        regions["cn-east-2"] = Region("cn-east-2","https://evs.cn-east-2.myhuaweicloud.com") ;
-        regions["cn-east-3"] = Region("cn-east-3","https://evs.cn-east-3.myhuaweicloud.com") ;
-        regions["cn-south-1"] = Region("cn-south-1","https://evs.cn-south-1.myhuaweicloud.com") ;
-        regions["cn-south-2"] = Region("cn-south-2","https://evs.cn-south-2.myhuaweicloud.com") ;
-        regions["cn-southwest-2"] = Region("cn-southwest-2","https://evs.cn-southwest-2.myhuaweicloud.com") ;
-        regions["ap-southeast-2"] = Region("ap-southeast-2","https://evs.ap-southeast-2.myhuaweicloud.com") ;
-        regions["ap-southeast-1"] = Region("ap-southeast-1","https://evs.ap-southeast-1.myhuaweicloud.com") ;
-        regions["ap-southeast-3"] = Region("ap-southeast-3","https://evs.ap-southeast-3.myhuaweicloud.com") ;
-        regions["sa-brazil-1"] = Region("sa-brazil-1","https://evs.sa-brazil-1.myhuaweicloud.com") ;
-        regions["la-north-2"] = Region("la-north-2","https://evs.la-north-2.myhuaweicloud.com") ;
-        regions["la-south-2"] = Region("la-south-2","https://evs.la-south-2.myhuaweicloud.com") ;
-        regions["na-mexico-1"] = Region("na-mexico-1","https://evs.na-mexico-1.myhuaweicloud.com") ;
-        regions["cn-north-9"] = Region("cn-north-9","https://evs.cn-north-9.myhuaweicloud.com") ;
-        regions["ap-southeast-4"] = Region("ap-southeast-4","https://evs.ap-southeast-4.myhuaweicloud.com") ;
+        regions["af-south-1"] = Region("af-south-1",
+            { 
+                "https://evs.af-south-1.myhuaweicloud.com"
+            });
+        regions["cn-north-4"] = Region("cn-north-4",
+            { 
+                "https://evs.cn-north-4.myhuaweicloud.com"
+            });
+        regions["cn-north-1"] = Region("cn-north-1",
+            { 
+                "https://evs.cn-north-1.myhuaweicloud.com"
+            });
+        regions["cn-north-2"] = Region("cn-north-2",
+            { 
+                "https://evs.cn-north-2.myhuaweicloud.com"
+            });
+        regions["cn-east-2"] = Region("cn-east-2",
+            { 
+                "https://evs.cn-east-2.myhuaweicloud.com"
+            });
+        regions["cn-east-3"] = Region("cn-east-3",
+            { 
+                "https://evs.cn-east-3.myhuaweicloud.com"
+            });
+        regions["cn-south-1"] = Region("cn-south-1",
+            { 
+                "https://evs.cn-south-1.myhuaweicloud.com"
+            });
+        regions["cn-south-2"] = Region("cn-south-2",
+            { 
+                "https://evs.cn-south-2.myhuaweicloud.com"
+            });
+        regions["cn-southwest-2"] = Region("cn-southwest-2",
+            { 
+                "https://evs.cn-southwest-2.myhuaweicloud.com"
+            });
+        regions["ap-southeast-2"] = Region("ap-southeast-2",
+            { 
+                "https://evs.ap-southeast-2.myhuaweicloud.com"
+            });
+        regions["ap-southeast-1"] = Region("ap-southeast-1",
+            { 
+                "https://evs.ap-southeast-1.myhuaweicloud.com"
+            });
+        regions["ap-southeast-3"] = Region("ap-southeast-3",
+            { 
+                "https://evs.ap-southeast-3.myhuaweicloud.com"
+            });
+        regions["sa-brazil-1"] = Region("sa-brazil-1",
+            { 
+                "https://evs.sa-brazil-1.myhuaweicloud.com"
+            });
+        regions["la-north-2"] = Region("la-north-2",
+            { 
+                "https://evs.la-north-2.myhuaweicloud.com"
+            });
+        regions["la-south-2"] = Region("la-south-2",
+            { 
+                "https://evs.la-south-2.myhuaweicloud.com"
+            });
+        regions["na-mexico-1"] = Region("na-mexico-1",
+            { 
+                "https://evs.na-mexico-1.myhuaweicloud.com"
+            });
+        regions["cn-north-9"] = Region("cn-north-9",
+            { 
+                "https://evs.cn-north-9.myhuaweicloud.com"
+            });
+        regions["ap-southeast-4"] = Region("ap-southeast-4",
+            { 
+                "https://evs.ap-southeast-4.myhuaweicloud.com"
+            });
         return regions;
     }
 private: