+++ frs/include/huaweicloud/frs/v2/FrsRegion.h
@@ -22,12 +22,30 @@ public:
 
     static std::map<std::string, Region> initRegions() {
         std::map<std::string, Region> regions;
-        regions["cn-north-1"] = Region("cn-north-1","https://face.cn-north-1.myhuaweicloud.com") ;
-        regions["cn-north-4"] = Region("cn-north-4","https://face.cn-north-4.myhuaweicloud.com") ;
-        regions["cn-south-1"] = Region("cn-south-1","https://face.cn-south-1.myhuaweicloud.com") ;
-        regions["cn-east-3"] = Region("cn-east-3","https://face.cn-east-3.myhuaweicloud.com") ;
-        regions["ap-southeast-1"] = Region("ap-southeast-1","https://face.ap-southeast-1.myhuaweicloud.com") ;
-        regions["ap-southeast-2"] = Region("ap-southeast-2","https://face.ap-southeast-2.myhuaweicloud.com") ;
+        regions["cn-north-1"] = Region("cn-north-1",
+            { 
+                "https://face.cn-north-1.myhuaweicloud.com"
+            });
+        regions["cn-north-4"] = Region("cn-north-4",
+            { 
+                "https://face.cn-north-4.myhuaweicloud.com"
+            });
+        regions["cn-south-1"] = Region("cn-south-1",
+            { 
+                "https://face.cn-south-1.myhuaweicloud.com"
+            });
+        regions["cn-east-3"] = Region("cn-east-3",
+            { 
+                "https://face.cn-east-3.myhuaweicloud.com"
+            });
+        regions["ap-southeast-1"] = Region("ap-southeast-1",
+            { 
+                "https://face.ap-southeast-1.myhuaweicloud.com"
+            });
+        regions["ap-southeast-2"] = Region("ap-southeast-2",
+            { 
+                "https://face.ap-southeast-2.myhuaweicloud.com"
+            });
         return regions;
     }
 private: