+++ sis/include/huaweicloud/sis/v1/SisRegion.h
@@ -22,9 +22,18 @@ public:
 
     static std::map<std::string, Region> initRegions() {
         std::map<std::string, Region> regions;
-        regions["cn-north-1"] = Region("cn-north-1","https://sis-ext.cn-north-1.myhuaweicloud.com") ;
-        regions["cn-north-4"] = Region("cn-north-4","https://sis-ext.cn-north-4.myhuaweicloud.com") ;
-        regions["cn-east-3"] = Region("cn-east-3","https://sis-ext.cn-east-3.myhuaweicloud.com") ;
+        regions["cn-north-1"] = Region("cn-north-1",
+            { 
+                "https://sis-ext.cn-north-1.myhuaweicloud.com"
+            });
+        regions["cn-north-4"] = Region("cn-north-4",
+            { 
+                "https://sis-ext.cn-north-4.myhuaweicloud.com"
+            });
+        regions["cn-east-3"] = Region("cn-east-3",
+            { 
+                "https://sis-ext.cn-east-3.myhuaweicloud.com"
+            });
         return regions;
     }
 private: